+++ src/objects/H2DE_bar_object.cpp
@@ -1,7 +1,7 @@
-#include "H2DE/H2DE_bar_object.h"
+#include "H2DE/H2DE_object.h"
 
 // INIT
-H2DE_BarObject::H2DE_BarObject(H2DE_Engine* engine, H2DE_ObjectData od, H2DE_BarObjectData b) : H2DE_Object(engine, od), bod(b), value(b.defaultValue) {
+H2DE_BarObject::H2DE_BarObject(H2DE_Engine* engine, H2DE_ObjectData od, H2DE_BarObjectData b) : H2DE_Object(engine, od), bod(b), percentage(b.defaultPercentage) {
 
 }
 
@@ -22,10 +22,10 @@ H2DE_BarObject::~H2DE_BarObject() {
 }
 
 // UPDATE
-void H2DE_BarObject::updateImpl() {
-    std::cout << value << std::endl;
-
-    // bod.front->sd.size // je sais aps trop comment faire pour changer la size de la bar front sans changer la size du background
+void H2DE_BarObject::update() {
+    if (bod.front) {
+        bod.front->percentage = percentage;
+    }
 }
 
 // GETTER
@@ -33,7 +33,15 @@ std::vector<H2DE_Surface*> H2DE_BarObject::getSurfaces() const {
     return { bod.background, bod.front };
 }
 
+H2DE_Surface* H2DE_GetBarFrontSurface(const H2DE_BarObject* bar) {
+    return bar->bod.front;
+}
+
+H2DE_Surface* H2DE_GetBarBackgroundSurface(const H2DE_BarObject* bar) {
+    return bar->bod.background;
+}
+
 // SETTER
-void H2DE_SetBarValue(H2DE_Engine* engine, H2DE_BarObject* bar, float value) {
-    bar->value = value;
+void H2DE_SetBarPercentage(H2DE_BarObject* bar, float percentage) {
+    bar->percentage = std::clamp(percentage, 0.0f, 100.0f);
 }