+++ src/objects/H2DE_button_object.cpp
@@ -1,4 +1,4 @@
-#include "H2DE/H2DE_button_object.h"
+#include "H2DE/H2DE_object.h"
 
 // INIT
 H2DE_ButtonObject::H2DE_ButtonObject(H2DE_Engine* engine, H2DE_ObjectData od, H2DE_ButtonObjectData b) : H2DE_Object(engine, od), bod(b) {
@@ -19,7 +19,7 @@ H2DE_ButtonObject::~H2DE_ButtonObject() {
 }
 
 // UPDATE
-void H2DE_ButtonObject::updateImpl() {
+void H2DE_ButtonObject::update() {
     if (!bod.onclick && !bod.onhover && !bod.onout) {
         return;
     }