+++ src/surface/H2DE_sprite.cpp
@@ -1,14 +1,22 @@
-#include "H2DE/H2DE_sprite.h"
+#include "H2DE/H2DE_surface.h"
 
 // INIT
 H2DE_Sprite::H2DE_Sprite(H2DE_Engine* engine, const H2DE_SurfaceData& sd, const H2DE_SpriteData& s) : H2DE_Surface(engine, sd), spd(s) {
-    H2DE_Delay(engine, spd.delay, [this]() {
-        nextFrame();
-    }, -1, spd.pauseSensitive);
+    initDelay();
 }
 
 H2DE_Surface* H2DE_CreateSprite(H2DE_Engine* engine, const H2DE_SurfaceData& sd, const H2DE_SpriteData& spd) {
     return new H2DE_Sprite(engine, sd, spd);
+}
+
+void H2DE_Sprite::initDelay() {
+    if (delayId != -1) {
+        H2DE_StopDelay(engine, delayId, false);
+    }
+
+    delayId = H2DE_Delay(engine, spd.delay, [this]() {
+        nextFrame();
+    }, -1, spd.pauseSensitive);
 }
 
 // CLEANUP
@@ -33,4 +41,54 @@ std::optional<H2DE_AbsRect> H2DE_Sprite::getSrcRect() const {
     srcRect.addSize(spd.size);
 
     return srcRect;
+}
+
+// SETTER
+void H2DE_SetSpriteStartingPos(H2DE_Surface* sprite, const H2DE_AbsPos& startingPos) {
+    H2DE_Sprite* spr = dynamic_cast<H2DE_Sprite*>(sprite);
+    
+    if (spr) {
+        spr->spd.startingPos = startingPos;
+    }
+}
+
+void H2DE_SetSpriteSize(H2DE_Surface* sprite, const H2DE_AbsSize& size) {
+    H2DE_Sprite* spr = dynamic_cast<H2DE_Sprite*>(sprite);
+    
+    if (spr) {
+        spr->spd.size = size;
+    }
+}
+
+void H2DE_SetSpriteSpacing(H2DE_Surface* sprite, int spacing) {
+    H2DE_Sprite* spr = dynamic_cast<H2DE_Sprite*>(sprite);
+    
+    if (spr) {
+        spr->spd.spacing = spacing;
+    }
+}
+
+void H2DE_SetSpriteNbFrame(H2DE_Surface* sprite, unsigned int nbFrame) {
+    H2DE_Sprite* spr = dynamic_cast<H2DE_Sprite*>(sprite);
+    
+    if (spr) {
+        spr->spd.nbFrame = nbFrame;
+    }
+}
+
+void H2DE_SetSpriteDelay(H2DE_Surface* sprite, unsigned int delay) {
+    H2DE_Sprite* spr = dynamic_cast<H2DE_Sprite*>(sprite);
+    
+    if (spr) {
+        spr->spd.delay = delay;
+        spr->initDelay();
+    }
+}
+
+void H2DE_SetSpritePauseSensitive(H2DE_Surface* sprite, bool pauseSensitive) {
+    H2DE_Sprite* spr = dynamic_cast<H2DE_Sprite*>(sprite);
+    
+    if (spr) {
+        spr->spd.pauseSensitive = pauseSensitive;
+    }
 }