+++ src/surface/H2DE_texture.cpp
@@ -1,4 +1,4 @@
-#include "H2DE/H2DE_texture.h"
+#include "H2DE/H2DE_surface.h"
 
 // INIT
 H2DE_Texture::H2DE_Texture(H2DE_Engine* engine, const H2DE_SurfaceData& sd, const H2DE_TextureData& t) : H2DE_Surface(engine, sd), ted(t) {
@@ -17,4 +17,13 @@ H2DE_Texture::~H2DE_Texture() {
 // GETTER
 std::optional<H2DE_AbsRect> H2DE_Texture::getSrcRect() const {
     return ted.srcRect;
+}
+
+// SETTER
+void H2DE_SetTextureSrcRect(H2DE_Surface* texture, const std::optional<H2DE_AbsRect>& srcRect) {
+    H2DE_Texture* tex = dynamic_cast<H2DE_Texture*>(texture);
+    
+    if (tex) {
+        tex->ted.srcRect = srcRect;
+    }
 }